Soften the check of AASConstraint 108
Before on adding new element into ´SubmodelElementList´ it checks if the added element is of type `SubmodelElementList.value_type_list_element`. This check does not allow to create custom classes of SubmodelElements. See #47
That is why we soften it now and check if new objects are instances of the given type.

#47: https://github.com/eclipse-basyx/basyx-python-sdk/issues/47